+++ typeddfs/cli_help.py
@@ -10,7 +10,6 @@ Utils for getting nice CLI help on DataFrame inputs.
     reStructuredText, and `Typer <https://github.com/tiangolo/typer>`_.
 """
 from dataclasses import dataclass
-from inspect import cleandoc
 from typing import FrozenSet, Mapping, Sequence, Type
 
 from typeddfs import TypedDf, MatrixDf
@@ -32,13 +31,27 @@ class DfFormatHelp:
     fmt: FileFormat
     desc: str
 
+    @property
+    def bare_suffixes(self) -> Sequence[str]:
+        """
+        Returns all suffixes, excluding compressed variants (etc. ``.gz``), naturally sorted.
+        """
+        suffixes = {CompressionFormat.strip_suffix(s).name for s in self.fmt.suffixes}
+        return Utils.natsort(suffixes, str)
+
+    @property
+    def all_suffixes(self) -> Sequence[str]:
+        """
+        Returns all suffixes, naturally sorted.
+        """
+        return Utils.natsort(self.fmt.suffixes, str)
+
     def get_text(self) -> str:
         """
         Returns a 1-line string of the suffixes and format description.
         """
-        suffixes = Utils.natsort(self.fmt.suffixes, str)
+        suffixes = self.bare_suffixes
         if self.fmt.is_text:
-            suffixes = {CompressionFormat.strip_suffix(s).name for s in suffixes}
             comps = ["." + c.name for c in CompressionFormat.list_non_empty()]
             return "/".join(suffixes) + "[" + "/".join(comps) + "]" + ": " + self.desc
         else:
@@ -54,28 +67,48 @@ class DfFormatsHelp(FrozenSet[DfFormatHelp]):
         """
         Returns a single-line text listing of allowed file formats.
 
+        Args:
+            recommended_only: Skip non-recommended file formats
+
         Returns:
             Something like::
-                .csv, .tsv/.tab, .flexwf [optionally .gz, .xz, .zip, or .bz2], .feather, .snappy, ...
+                .csv, .tsv/.tab, or .flexwf [.gz,/.xz,/.zip/.bz2]; .feather, .pickle, or .snappy ...
         """
         fmts = [f for f in self if not recommended_only or f.fmt.is_recommended]
-        text_fmts = Utils.natsort([f.get_text() for f in fmts if f.fmt.is_text], dtype=str)
-        bin_fmts = Utils.natsort([f.get_text() for f in fmts if f.fmt.is_binary], dtype=str)
+        text_fmts = Utils.natsort(
+            ["/".join(f.bare_suffixes) for f in fmts if f.fmt.is_text], dtype=str
+        )
+        bin_fmts = Utils.natsort(
+            ["/".join(f.bare_suffixes) for f in fmts if f.fmt.is_binary], dtype=str
+        )
         txt = ""
         if len(text_fmts) > 0:
             txt += (
-                ", ".join(text_fmts)
-                + " [optionally "
-                + ", ".join([s.name for s in CompressionFormat.list_non_empty()])
+                Utils.join_to_str(*text_fmts, last="or")
+                + " ["
+                + "/".join([s.suffix for s in CompressionFormat.list_non_empty()])
                 + "]"
             )
         if len(bin_fmts) > 0:
-            txt += (", " if len(text_fmts) > 0 else "") + ", ".join(bin_fmts)
+            txt += ("; " if len(text_fmts) > 0 else "") + Utils.join_to_str(*bin_fmts, last="or")
         return txt
 
-    def get_long_text(self, *, recommended_only: bool = False) -> str:
+    def get_long_text(
+        self,
+        *,
+        recommended_only: bool = False,
+        nl: str = "\n",
+        bullet: str = "- ",
+        indent: str = "  ",
+    ) -> str:
         """
         Returns a multi-line text listing of allowed file formats.
+
+        Args:
+            recommended_only: Skip non-recommended file formats
+            nl: Newline characters; use "\\n", "\\\\n", or " "
+            bullet: Prepended to each item
+            indent: Spaces for nested indent
 
         Returns:
             Something like::
@@ -89,10 +122,9 @@ class DfFormatsHelp(FrozenSet[DfFormatHelp]):
 
                 .pickle/.pkl: Python Pickle [discouraged]
         """
-        nl = "\n\n"
-        bullet = nl + " " * 2 + "- "
+        bullet = nl + indent + bullet
         fmts = [f for f in self if not recommended_only or f.fmt.is_recommended]
-        formats = [f.get_text() + ("" if f.fmt.is_recommended else " [discouraged]") for f in fmts]
+        formats = [f.get_text() + ("" if f.fmt.is_recommended else " [avoid]") for f in fmts]
         formats = Utils.natsort(formats, str)
         txt = bullet + bullet.join(formats)
         return f"[[ Supported formats ]]: {txt}"
@@ -112,29 +144,45 @@ class DfHelp:
         return self.clazz.get_typing()
 
     def get_short_typing_text(self) -> str:
+        """
+        Returns 1-line text on only the required columns / structure.
+        """
         raise NotImplementedError()
 
     def get_long_typing_text(self) -> str:
+        """
+        Returns multi-line text on only the required columns / structure.
+        """
         raise NotImplementedError()
 
-    def get_short_text(self, *, use_doc: bool = True, recommended_only: bool = False) -> str:
+    def get_short_text(
+        self, *, use_doc: bool = True, recommended_only: bool = False, nl: str = "\n"
+    ) -> str:
         """
         Returns a multi-line description with compressed text.
 
         Args:
             use_doc: Include the docstring of the DataFrame type
             recommended_only: Only include recommended formats
+            nl: Newline characters; use "\\n", "\\\\n", or " "
         """
-        nl = "\n\n"
         t = self.get_short_typing_text()
         return (
             self.get_header_text(use_doc=use_doc)
             + ((nl + t) if len(t) > 0 else "")
             + nl
-            + self.formats.get_short_text()
+            + self.formats.get_short_text(recommended_only=recommended_only)
         ).replace(nl * 2, nl)
 
-    def get_long_text(self, *, use_doc: bool = True, recommended_only: bool = False) -> str:
+    def get_long_text(
+        self,
+        *,
+        use_doc: bool = True,
+        recommended_only: bool = False,
+        nl: str = "\n",
+        bullet: str = "- ",
+        indent: str = "  ",
+    ) -> str:
         """
         Returns a multi-line text description of the DataFrame.
         Includes its required and optional columns, and supported file formats.
@@ -142,22 +190,27 @@ class DfHelp:
         Args:
             use_doc: Include the docstring of the DataFrame type
             recommended_only: Only include recommended formats
+            nl: Newline characters; use "\\n", "\\n\\n", or " "
+            bullet: Prepended to each item
+            indent: Spaces for nested indent
         """
-        nl = "\n\n"
         t = self.get_long_typing_text()
         return (
             self.get_header_text(use_doc=use_doc)
             + ((nl + t) if len(t) > 0 else "")
             + nl
-            + self.formats.get_long_text()
+            + self.formats.get_long_text(
+                recommended_only=recommended_only, nl=nl, bullet=bullet, indent=indent
+            )
         ).replace(nl * 2, nl)
 
-    def get_header_text(self, *, use_doc: bool = True) -> str:
+    def get_header_text(self, *, use_doc: bool = True, nl: str = "\n") -> str:
         """
         Returns a multi-line header of the DataFrame name and docstring.
 
         Args:
             use_doc: Include the docstring, as long as it is not ``None``
+            nl: Newline characters; use "\\n", "\\n\\n", or " "
 
         Returns:
             Something like::
@@ -165,7 +218,6 @@ class DfHelp:
 
                 This is a big table for big things.
         """
-        nl = "\n\n"
         s = f"A {self.clazz.__name__} file."
         if use_doc and self.clazz.__doc__ is not None:
             s += nl + self.clazz.__doc__
@@ -177,31 +229,47 @@ class TypedDfHelp(DfHelp):
         """
         Returns a condensed text description of the required and optional columns.
         """
+        t = self.typing
+        req = self.get_required_cols(short=True)
+        res = self.get_reserved_cols(short=True)
         s = ""
-        if len(self.required_cols) > 0:
-            s += f"Requires columns: {', '.join(self.required_cols)}."
-        if len(self.reserved_cols) > 0 and self.typing.is_strict:
-            s += (" " if len(s) > 0 else " ") + f"Permits columns: {', '.join(self.reserved_cols)}"
-        elif len(self.reserved_cols) > 0:
+        if len(req) > 0:
+            s += f"Requires columns {Utils.join_to_str(*req, last='and')}."
+        if len(res) > 0:
             s += (
                 (" " if len(s) > 0 else " ")
-                + f"Recognizes columns: {', '.join(self.reserved_cols)}. Additional columns are ignored."
+                + "Columns "
+                + Utils.join_to_str(*res, last="and")
+                + " are optional."
             )
+        s += " "
+        if t.is_strict:
+            s += "More columns are ok."
+        else:
+            s += "No extra columns are allowed."
         return s
 
-    def get_long_typing_text(self) -> str:
+    def get_long_typing_text(
+        self, *, nl: str = "\n", bullet: str = "- ", indent: str = "  "
+    ) -> str:
         """
         Returns a long text description of the required and optional columns.
+
+        Args:
+            nl: Newline characters; use "\\n", "\\n\\n", or " "
+            bullet: Prepended to each item
+            indent: Spaces for nested indent
         """
-        nl = "\n\n"
-        bullet = nl + " " * 2 + "- "
+        bullet = nl + indent + bullet
+        req = self.get_required_cols(short=False)
+        res = self.get_reserved_cols(short=False)
         s = ""
-        if len(self.required_cols) > 0:
-            s += f"[[ Required columns ]]: {bullet}{bullet.join(self.required_cols)}"
-        if len(self.reserved_cols) > 0:
+        if len(req) > 0:
+            s += f"[[ Required columns ]]: {bullet}{bullet.join(req)}"
+        if len(res) > 0:
             if len(s) > 0:
                 s += nl
-            s += f"[[ Optional columns ]]: {bullet}{bullet.join(self.reserved_cols)}"
+            s += f"[[ Optional columns ]]: {bullet}{bullet.join(res)}"
         if len(s) == 0:
             return s
         if not self.typing.is_strict:
@@ -210,25 +278,29 @@ class TypedDfHelp(DfHelp):
             s += f"{nl}No additional columns are allowed."
         return s
 
-    @property
-    def required_cols(self) -> Sequence[str]:
+    def get_required_cols(self, *, short: bool = False) -> Sequence[str]:
         """
         Lists required columns and their data types.
-        """
-        return self._cols(self.typing.required_names)
 
-    @property
-    def reserved_cols(self) -> Sequence[str]:
+        Args:
+            short: Use shorter strings (e.g. "int" instead of "integer")
+        """
+        return self._cols(self.typing.required_names, short=short)
+
+    def get_reserved_cols(self, *, short: bool = False) -> Sequence[str]:
         """
         Lists reserved (optional) columns and their data types.
-        """
-        return self._cols(self.typing.reserved_names)
 
-    def _cols(self, which: Sequence[str]) -> Sequence[str]:
+        Args:
+            short: Use shorter strings (e.g. "int" instead of "integer")
+        """
+        return self._cols(self.typing.reserved_names, short=short)
+
+    def _cols(self, which: Sequence[str], *, short: bool) -> Sequence[str]:
         lst = []
         for c in which:
             t = self.typing.auto_dtypes.get(c)
-            t = Utils.describe_dtype(t)
+            t = Utils.describe_dtype(t, short=short)
             if t is None:
                 lst.append(c)
             else:
@@ -243,21 +315,22 @@ class MatrixDfHelp(DfHelp):
         """
         t = self.typing
         if t.value_dtype is None:
-            s = "Numeric matrix. "
+            s = "Matrix. "
         else:
-            s = f"{t.value_dtype.__name__} matrix. "
+            s = Utils.describe_dtype(t.value_dtype).capitalize()
+            s += f" ({t.value_dtype.__name__}) matrix. "
         s += f"List row names in the index or a special column 'row'."
         return s
 
-    def get_long_typing_text(self) -> str:
+    def get_long_typing_text(self, *, nl: str = " ") -> str:
         """
         Returns a long text description of the required format for a matrix.
         """
         t = self.typing
-        s = "Numeric matrix with named rows and columns. "
-        s += f"List row names in the index or a special column 'row'. "
+        s = "Numeric matrix with named (string-typed) rows and columns." + nl
+        s += f"List row names in the index or a special column 'row'."
         if t.value_dtype is not None:
-            s += f"Values are cast to {t.value_dtype.__name__}"
+            s += nl + f"Values are cast to {t.value_dtype.__name__}"
         return s